+++ Demo/Demo/SceneDelegate.swift
@@ -34,7 +34,11 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 }
 
 extension SceneDelegate: TurboNavigationDelegate {
-    func session(_ session: Turbo.Session, didFailRequestForVisitable visitable: Turbo.Visitable, error: Error) {
-        print("An error occurred loading a visit:", error)
+    func session(_ session: Session, didFailRequestForVisitable visitable: Visitable, error: Error) {
+        if let errorPresenter = visitable as? ErrorPresenter {
+            errorPresenter.presentError(error) {
+                session.reload()
+            }
+        }
     }
 }